Twitch Chat Enhancements:
Allowing multiple messages per Thread loop.  Before it could on process one command per 250 ms, now it will process as many commands are in the queue

Moved to ArrayBlockingQueue
No need for apache collections dependency
Still FIFO, performs better, native Java object
Allows to wait for an object to become available in the queue on a get call.  Now we wait for 1 second before Thread continues. This also removes the need to check the queue size which can be cause locking

Two new functions to check if the bot has joined a channel already and get list of active channels